Correct comments to improve documentation
The auto-generated documentation pulls in the contents of various
'Example' functions, including comments. When looking at the docs, I
noticed that a comment in ExampleSegmentList_SetExif() incorrectly
claimed to be updating the 'UserComment' tag, likely due to being
copy/pasted from another function that does update this tag.

Here, correct the comments in both versions of
ExampleSegmentList_SetExif() to list the tags they are actually
updating.

For TestSegmentList_ConstructExifBuilder(t *testing.T) remove the
'Update the UserComment tag' comment entirely, as this function does not
update any tags.